feat: add list/domains command
