+++ messages/list/domains/messages.go
@@ -1,0 +1,8 @@
+package domains
+
+var (
+	DomainsListUsage            = "domains"
+	DomainsListShortDescription = "Displays your domains"
+	DomainsListLongDescription  = "Displays all your domain references to your edge locations"
+	DomainsListHelpFlag         = "Displays more information about the list domain command"
+)
