+++ pkg/cmd/list/domains/domains_test.go
@@ -1,4 +1,4 @@
-package list
+package domains
 
 import (
 	"testing"
