+++ pkg/cmd/list/domains/fixtures/domains.json
