+++ pkg/cmd/list/domains/fixtures/nodomains.json
