+++ pkg/cmd/list/list.go
@@ -3,6 +3,7 @@ package list
 import (
 	"github.com/MakeNowJust/heredoc"
 	msg "github.com/aziontech/azion-cli/messages/list"
+	domains "github.com/aziontech/azion-cli/pkg/cmd/list/domains"
 	edgeApplications "github.com/aziontech/azion-cli/pkg/cmd/list/edge_applications"
 	rule "github.com/aziontech/azion-cli/pkg/cmd/list/rule_engine"
 
@@ -25,6 +26,7 @@ func NewCmd(f *cmdutil.Factory) *cobra.Command {
 
 	cmd.AddCommand(edgeApplications.NewCmd(f))
 	cmd.AddCommand(rule.NewCmd(f))
+	cmd.AddCommand(domains.NewCmd(f))
 
 	cmd.Flags().BoolP("help", "h", false, msg.FlagHelp)
 	return cmd
